+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Replace integration tests by visual regression tests
 - Fix access permissions in PDF version 1.7ext3
 - Fix Buffer() is deprecation warning
+- Fix "@" in FontName
 
 ### [v0.11.0] - 2019-12-03