+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ### Unreleased
 
+- Add support for Embedded Files and File Attachment Annotations
+- Accessibility support
 - Replace integration tests by visual regression tests
 - Fix access permissions in PDF version 1.7ext3
 - Fix Buffer() is deprecation warning
 - Add `forms.md` to generate documentation files
+- Fix "@" in FontName
+
 
 ### [v0.11.0] - 2019-12-03