+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 - Replace integration tests by visual regression tests
 - Fix access permissions in PDF version 1.7ext3
 - Fix Buffer() is deprecation warning
+- Add `forms.md` to generate documentation files
 - Fix "@" in FontName
+
 
 ### [v0.11.0] - 2019-12-03